MIA Algorihtms v.1.1.2 Simulated Annealing Phrasing v.1.0.2
Simulated Annealing corrected: Before when it corrects the size of an exceeded phrase it used to put as silence genuine original notes. Also corrected some errors on this re-size method. TODO: Triplets cause uncorrectable errors sometimes when they appear: 1. The [_neighbor] method can cause triplets to appear in quantities other than three, and 2. In cases where sentences that are too long need to be corrected, where the last syllable is a triplet, I don't know what to do