MIA Harmfirst b.0.3.0
MIA first model [Harmfirst: Harmony first]: b.0.2.0 but with MIDI integration + fixes on Rythm for chords and bars adjustement. It throws correctly MIDI files and TXT with melody and harmony integration, repiting in the MIDI to times and with song metadata. Also it has an extra function for Progressions in which it uses silences if the chord maximun space figure is not enough for fill the phrase space

TODO: 1. Correct the octavation on the chords generation; 2. Integrate the sentences generation; 3. Create melodic tuning for Melody; 4. Building of MIA Harmfirst v1 complete architecture