MIA Harmfirst b.0.3.1
MIA first model [Harmfirst: Harmony first]: b.0.3.0 but with fixed error. Last version of rythm make adjustement of phrases if they exceeds their used size of the bars, but just changing the last syllable, there are some cases in which it is not enough, and it is necessary to set more syllables until the phrase fit completly. Using a loop it ensures this situation in this fix error (specially for chords)